add poolbuilder auto cleanup issues on kg web site http11

+++ Echoes/Clients/H2/HeaderEncoder.cs
@@ -30,8 +30,10 @@ namespace Echoes.H2
             Span<byte> buffer = stackalloc byte[_streamSetting.Remote.MaxHeaderLine];
             var encodedHeader = _hpackEncoder.Encode(encodingJob.Data, buffer);
 
-            var res = Packetizer.PacketizeHeader(encodedHeader, destinationBuffer.Span, endStream, encodingJob.StreamIdentifier,
-                (int) _streamSetting.Remote.MaxFrameSize, encodingJob.StreamDependency);
+            var res = Packetizer.PacketizeHeader(
+                encodedHeader, destinationBuffer.Span, 
+                endStream, encodingJob.StreamIdentifier,
+                _streamSetting.Remote.MaxFrameSize, encodingJob.StreamDependency);
 
             return destinationBuffer.Slice(0, res.Length);
         }
